// src/constants/env.type.ts
export enum EEnvType {
  APP_URL = 'APP_URL',
  NODE_ENV = 'NODE_ENV',
  PORT = 'PORT',
  DB_HOST = 'DB_HOST',
  DB_PORT = 'DB_PORT',
  DB_USERNAME = 'DB_USERNAME',
  DB_PASSWORD = 'DB_PASSWORD',
  DB_NAME = 'DB_NAME',
  PGADMIN_PORT = 'PGADMIN_PORT',
  REDIS_PASSWORD = 'REDIS_PASSWORD',
  STRIPE_API_KEY = 'STRIPE_API_KEY',
  LINE_CHANNEL_ACCESS_TOKEN = 'LINE_CHANNEL_ACCESS_TOKEN',
  LINE_CHANNEL_SECRET = 'LINE_CHANNEL_SECRET',
  JWT_SECRET_KEY = 'JWT_SECRET_KEY',
  S3_ACCESS_KEY = 'S3_ACCESS_KEY',
  S3_SECRET_KEY = 'S3_SECRET_KEY',
  S3_BUCKET = 'S3_BUCKET',
  S3_REGION = 'S3_REGION',
  END_POINT_MEDIA = 'END_POINT_MEDIA',
  BCRYPT_SALT_ROUNDS = 'BCRYPT_SALT_ROUNDS',
  SMTP_HOST = 'SMTP_HOST',
  SMTP_PORT = 'SMTP_PORT',
  SMTP_SECURE = 'SMTP_SECURE',
  SMTP_USER = 'SMTP_USER',
  SMTP_PASS = 'SMTP_PASS',
  MAIL_FROM = 'MAIL_FROM',
}